bugfix: deadlock when reading state of "Google Location Accuracy" toggle
State of this toggle is read from ContentProvider of Google Services Framework, which might not be
running when the check is performed.

If it's not running, OS spawns the process of this ContentProvider, which unconditionally connects
to GmsCompat app during initialization.
Part of the connection process happens on the main thread of GmsCompat app, which is deadlocked by
waiting for ContentProvider call to return.

Previously, this wasn't an issue because GmsCompat UI ran in a separate process.

Move connection to GSF ContentProvider to a background thread to avoid the deadlock.

Signed-off-by: Dmitrii <bankersenator@gmail.com>