notify the user about lack of POST_NOTIFICATIONS permission
Show a notification when a bound GMS component tries to show a notification, but doesn't have the
permission for it.

GmsCore doesn't have a good way to ask for this permission itself, which leads to the user missing
important notifications from it, eg about the need to make the device temporarily visible to nearby
devices to receive files through Nearby Share.

Signed-off-by: Dmitrii <bankersenator@gmail.com>