update binder redirection implementation
"Binder redirection" means replacement of destination of a subset of binder calls.

Before this commit, original destination was GmsCore and new destination was GmsCompatApp.

This commit (and its corresponding fw/base changes) replaces the new destination:
new destination is now the caller itself.

This means that redirected calls are no longer IPC, they happen inside the caller's process.

Binder redirection setup sequence is now as follows:
- caller encounters a binder that is eligible for redirection
- caller asks GmsCompatApp whether binder redirection should be used for that binder
- if yes, GmsCompatApp replies to caller with a BinderDef (binder definition).
BinderDef specifies which binder calls should be redirected, and how to construct a binder that will
be the new destination of these calls. Currently, it's constructed by loading code from GmsCompatApp
into caller process
- caller uses BinderDef to construct a binder inside caller's process and redirects specified calls
from the original destination to this new binder

This approach allows to significantly simplify and improve implementations of services that are
backed by these binders:
- no need to check for caller permissions
- no need to handle multiple callers
- no need to handle caller process dying
- OS correctly attributes battery and data usage (e.g. location data usage) to caller instead of
 attributing it to GmsCompatApp

Signed-off-by: Dmitrii <bankersenator@gmail.com>